draft cheer essay
I was on the fence about even storing this document anywhere, but didn't
want to risk losing the whole thing if I lost my laptop. I have a full
disk backup tool, but it's a bit heavy handed to recover a single
short markdown file.

So, here's the document, stored. I don't plan on publishing it. I've
copied it over to a private github gist, so... hopefully will never
lose it.